pyproject.toml
[project]
name = "crispy-meme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.43.0",
    "litellm>=1.80.5",
    "openai-agents[litellm]>=0.6.1",
]
